refactor(utils-vscode): relocate shapeFrom mapper into utils-vscode
